Do not attempt publish if PUBLISH_REPO is not set

+++ run.sh
@@ -1,5 +1,10 @@
 #!/usr/bin/env bash
 
+if [ -z $PUBLISH_REPO ]; then
+	echo "PUBLISH_REPO is not set. Exiting"
+	exit 0
+fi
+
 echo "Begin deploy."
 
 rm -rf deploy_repo || true
